requirements.txt
aiohttp

discord.py
python-dotenv
ollama